Create a table of all studies currently in data.frame

Create a table of all studies currently in data.frame

createStudyTable(bsdb.df, includeAlso = NULL)

Arguments

bsdb.df

data.frame produced by importBugSigDB, pre-filtered as desired

includeAlso

character with column names to additionally include in the output table (default = `NULL`)

Value

a data.frame of basic study information. Can be wrapped in kable_styling(kbl(...)) to format nicely.

Examples

full.dat <- bugsigdbr::importBugSigDB()
#> Using cached version from 2025-05-05 19:21:34
createStudyTable(full.dat)
#> # A tibble: 1,384 × 9
#>    `Study code` MaxCases MaxControls `Study design` Condition N_signatures PMID 
#>    <chr>           <dbl>       <dbl> <chr>          <chr>            <int> <chr>
#>  1 AartiDarra_…       54          54 cross-section… Hypergly…            8 NA   
#>  2 AartsE_2017        19          77 case-control   Attentio…            2 2886…
#>  3 AbangeWB_20…       87          82 cross-section… HIV infe…            3 3382…
#>  4 AbotsiRE_20…      174         174 randomized co… Response…            5 3680…
#>  5 Abrahamsson…       20          20 prospective c… Atopic e…            3 2215…
#>  6 AdambergK_2…        9           9 case-control   Obesity              6 2993…
#>  7 AdebayoAS_2…       46          25 case-control   Urinary …           14 2879…
#>  8 Adu-OppongB…       51          51 cross-section… Urinary …            7 3503…
#>  9 AguilarT_20…       16          56 cross-section… Obesity              2 3242…
#> 10 AhmedAlaa_2…       46          35 case-control   Parkinso…            8 NA   
#> # ℹ 1,374 more rows
#> # ℹ 2 more variables: DOI <chr>, URL <chr>