countBug counts the frequency of the most commonly identified bug in a simulated signature. getCriticalN performs a Monte Carlo simulation to estimate the number of times the most frequent taxon is expected to be observed in a list of signatures

countBug counts the frequency of the most commonly identified bug in a simulated signature. getCriticalN performs a Monte Carlo simulation to estimate the number of times the most frequent taxon is expected to be observed in a list of signatures

getCriticalN(relevant.sigs, siglengths, alpha = 0.05, nsim = 1000)

Arguments

relevant.sigs
a list of signatures that form the "background" from which taxa for simulated signatures will be drawn. These are used to estimate how frequently taxa occur

siglengths
The sizes of signatures found in a set of related studies. Simulated signatures will match these in number and size.

alpha
Probability at which a critical threshold will be calculated (default: 0.05)

nsim
Number of simulations (default: 1000)

Value

The 1 - alpha quantile of Monte Carlo simulated values for the maximum number of times any taxon is identified.

Details

E.g. for alpha = 0.05, we expect only a 5

Examples

full.dat <- importBugSigDB()

#> Error in importBugSigDB(): could not find function "importBugSigDB"

my.dat <- full.dat[full.dat$Curator == "Mst Afroza Parvin", ]

#> Error in eval(expr, envir, enclos): object 'full.dat' not found

relevant.dat <- full.dat[full.dat$`Body site` %in% my.dat$`Body site`, ]

#> Error in eval(expr, envir, enclos): object 'full.dat' not found

relevant.sigs <- extractBugSigs(my.dat)

#> Error in extractBugSigs(my.dat): object 'my.dat' not found

my.sigs.increased <- extractBugSigs(my.dat, direction = "increased")

#> Error in extractBugSigs(my.dat, direction = "increased"): object 'my.dat' not found

(my.siglengths <- sapply(my.sigs.increased, length))

#> Error in lapply(X = X, FUN = FUN, ...): object 'my.sigs.increased' not found

getCriticalN(relevant.sigs, my.siglengths)

#> Error in unlist(relevant.sigs): object 'relevant.sigs' not found

# Compare to observed
frequencySigs(my.sigs.increased)

#> Error in unlist(sigs): object 'my.sigs.increased' not found