Simulate a list of signatures based on a universe of taxa equal in size and length to some smaller set of signatures

Simulate a list of signatures based on a universe of taxa equal in size and length to some smaller set of signatures

simulateSignatures(relevant.sigs, siglengths)

Arguments

relevant.sigs
A list of signatures providing a relevant background for drawing taxa

siglengths
An integer vector, the length of which provides the number of signatures to be simulated, and the integers of which provide the number of taxa to be simulated in each signature

tax.level
Any subset of c("kingdom", "phylum", "class", "order", "family", "genus", "species", "strain")

Value

A list of signatures of the same number and individual lengths as found in my.dat

Examples

full.dat <- bugsigdbr::importBugSigDB()

#> Using cached version from 2021-04-23 16:14:17

my.dat <- full.dat[full.dat$Curator == "Mst Afroza Parvin", ]
relevant.dat <- full.dat[full.dat$`Body site` %in% my.dat$`Body site`, ]
relevant.sigs <- extractBugSigs(relevant.dat, tax.level = "genus")

#> Zero length signatures dropped:PMID:28038683_5 PMID:27015276_2 PMID:28467925_1 PMID:20140211_1 PMID:20140211_2 PMID:27228093_2 PMID:27228093_4 PMID:28636668_1 PMID:26261039_1 PMID:26261039_2 PMID:30051546_1 PMID:30051546_4 PMID:30051546_11 PMID:30051546_13 PMID:29922272_10 PMID:28442250_3 PMID:28442250_4 PMID:27450202_1 PMID:30541450_2 PMID:30541450_3 PMID:26683192_1 PMID:27001291_1 PMID:27001291_2 PMID:27001291_3 PMID:27001291_4 PMID:24223144_2 PMID:30656437_1 PMID:30656437_2 PMID:20844065_2 PMID:20844065_4 PMID:28018325_2 PMID:28018325_4 PMID:30237489_2 PMID:30237489_3 PMID:28222761_1 PMID:29950689_12 PMID:29388394_3 PMID:29388394_5 PMID:29388394_6 PMID:30228282_1 PMID:23526699_1 PMID:23526699_3 PMID:28035686_1 PMID:27306058_1 PMID:27306058_2 PMID:27306058_3 PMID:23032991_1 PMID:23032991_2 PMID:23032991_5 PMID:23032991_6 PMID:28512451_2 PMID:28512451_3 PMID:28512451_6 PMID:28512451_7 PMID:30224347_2 PMID:25764541_1 PMID:25764541_2 PMID:25764541_4 PMID:25764541_6 PMID:26919743_1 PMID:26919743_2 PMID:26919743_3 PMID:26919743_4 PMID:26919743_5 PMID:26919743_6 PMID:30867711_1 PMID:28628112_1 PMID:21829158_2 PMID:27741466_2 PMID:27741466_4 PMID:27741466_5 PMID:27741466_6 PMID:26428446_1 PMID:29596446_2 PMID:31053143_1 PMID:28624933_3 PMID:28624933_4 PMID:28624933_6 PMID:28624933_9 PMID:30644740_2 PMID:30644740_4 PMID:30644740_6 PMID:30644740_8 PMID:30188331_1 PMID:30188331_2 PMID:21885731_2 PMID:21885731_3 PMID:21885731_4 PMID:29738477_1 PMID:30081953_1 PMID:30081953_2 PMID:31595156_1 PMID:31595156_2 PMID:31595156_3 PMID:31595156_4 PMID:30791484_1 PMID:22339879_1 PMID:30764497_4 PMID:30764497_5 PMID:30707176_1 PMID:24532064_1 PMID:24532064_2 PMID:24532064_3 PMID:24532064_4 PMID:30355393_3 PMID:30355393_5 PMID:27292825_5 PMID:27812181_1 PMID:27812181_2 PMID:27812181_3 PMID:27812181_4 PMID:23339708_1 PMID:23339708_2 PMID:23339708_3 PMID:23339708_4 PMID:31528836_1 PMID:23038174_3 PMID:23038174_4 PMID:22153774_2 PMID:32009325_1 PMID:30139999_2 PMID:28538678_1 PMID:28538678_2 PMID:28538678_3 PMID:29624747_1 PMID:29624747_2 PMID:29624747_3 PMID:29624747_4 PMID:29624747_5 PMID:29624747_6 PMID:29624747_7 PMID:29624747_8 PMID:25599982_1 PMID:25599982_2 PMID:25599982_3 PMID:30418043_1 PMID:29971220_6 PMID:29971220_8 PMID:29971220_9 PMID:29971220_10 PMID:29167295_8 PMID:10202341_1 PMID:29434076_1 PMID:31289031_5 PMID:31289031_6 PMID:31289031_7 PMID:31289031_8 PMID:30838178_1 PMID:30838178_2 PMID:30838178_3 PMID:30838178_4 PMID:31678982_1 PMID:31678982_2 PMID:31678982_6 PMID:29740427_1 PMID:29740427_3 PMID:29740427_4 PMID:29740427_5 PMID:27176462_3 PMID:27176462_4 PMID:30674011_1 PMID:30674011_3 PMID:30674011_4 PMID:29518185_1 PMID:29518185_2 PMID:29518185_3 PMID:29518185_4 PMID:29518185_5 PMID:29518185_6 PMID:29518185_7 PMID:32466801_1 PMID:32466801_2 PMID:32466801_3 PMID:32466801_4 PMID:32115690_1 PMID:32115690_2 PMID:28714953_1 PMID:29186529_5 PMID:29783199_1 PMID:26971496_1 PMID:28789705_1 PMID:28789705_2 PMID:28789705_3 PMID:28789705_4 PMID:28789705_5 PMID:28789705_6 PMID:25444008_1 PMID:25444008_2 PMID:25444008_3 PMID:25444008_4 PMID:25444008_5 PMID:25444008_6 PMID:25444008_7 PMID:25444008_8 PMID:25444008_9 PMID:25444008_10 PMID:25444008_11 PMID:25444008_12 PMID:25444008_13 PMID:25444008_14 PMID:25444008_15 PMID:25444008_16 PMID:28733895_1 PMID:31006170_8 PMID:29097494_4 PMID:29097494_5 PMID:25319745_4 PMID:28903182_1 PMID:28903182_2 PMID:29554907_3 PMID:23638009_1 PMID:23638009_2 PMID:23638009_3 PMID:28808302_3 PMID:30568265_6 PMID:32056780_2 PMID:32056780_4 PMID:32056780_6 PMID:30267022_1 PMID:30267022_2 PMID:26551842_1 PMID:26551842_2 PMID:26551842_3 PMID:26551842_4 PMID:26551842_5 PMID:26551842_6 PMID:26551842_7 PMID:32143621_4 PMID:19849869_1 PMID:22546742_1 PMID:23631345_1 PMID:23631345_2 PMID:32429742_1 PMID:32429742_2 PMID:25761741_1 PMID:25761741_3 PMID:23459324_1 PMID:23459324_2 PMID:23459324_3 PMID:23459324_4 PMID:29935270_1 PMID:29935270_2 PMID:29935270_3 PMID:29935270_4 PMID:26745497_1 PMID:24748167_2 PMID:24748167_4 PMID:24748167_6 PMID:24748167_8 PMID:24748167_10 PMID:29592876_1 PMID:29592876_2 PMID:29592876_3 PMID:29592876_4 PMID:32065949_2 PMID:32065949_4 PMID:31001490_1 PMID:31001490_2 PMID:30314304_2 PMID:27640125_1 PMID:27640125_2 PMID:27640125_3 PMID:25944283_1 PMID:25944283_3 PMID:25944283_4 PMID:30860574_3 PMID:30860574_7 PMID:30860574_8 PMID:30860574_13 PMID:27634868_1 PMID:27634868_2 PMID:30944164_1 PMID:30944164_2 PMID:30944164_3 PMID:30944164_4 PMID:29874201_2 PMID:29874201_3 PMID:29874201_4 PMID:29874201_5 PMID:32747678_1 PMID:31727922_4 PMID:33294219_1 PMID:33294219_2 PMID:33294219_3 PMID:33294219_4 PMID:33294219_5 PMID:33294219_6 PMID:33294219_7 PMID:33294219_8 PMID:33294219_9 PMID:33294219_10 PMID:33294219_11 PMID:26416813_1 PMID:26416813_2 PMID:32690600_1 PMID:32690600_2

siglengths <- sapply(my.dat, length)
simulateSignatures(relevant.sigs, siglengths, tax.level = "genus")

#> Error in simulateSignatures(relevant.sigs, siglengths, tax.level = "genus"): unused argument (tax.level = "genus")