countBug counts the frequency of the most commonly identified bug in a simulated signature. getCriticalN performs a Monte Carlo simulation to estimate the number of times the most frequent taxon is expected to be observed in a list of signatures

countBug counts the frequency of the most commonly identified bug in a simulated signature. getCriticalN performs a Monte Carlo simulation to estimate the number of times the most frequent taxon is expected to be observed in a list of signatures

getCriticalN(relevant.sigs, siglengths, alpha = 0.05, nsim = 1000)

Arguments

relevant.sigs
a list of signatures that form the "background" from which taxa for simulated signatures will be drawn. These are used to estimate how frequently taxa occur

siglengths
The sizes of signatures found in a set of related studies. Simulated signatures will match these in number and size.

alpha
Probability at which a critical threshold will be calculated (default: 0.05)

nsim
Number of simulations (default: 1000)

Value

The 1 - alpha quantile of Monte Carlo simulated values for the maximum number of times any taxon is identified.

Details

E.g. for alpha = 0.05, we expect only a 5

Examples

full.dat <- bugsigdbr::importBugSigDB()

#> Using cached version from 2021-04-09 18:00:51

my.dat <- full.dat[full.dat$Curator == "Mst Afroza Parvin", ]
relevant.dat <- full.dat[full.dat$`Body site` %in% my.dat$`Body site`, ]
relevant.sigs <- extractBugSigs(my.dat)

#> Zero length signatures dropped:PMID:26745497_1

my.sigs.increased <- extractBugSigs(my.dat, direction = "increased")

#> Zero length signatures dropped:PMID:26745497_1

(my.siglengths <- sapply(my.sigs.increased, length))

#> PMID:27411009_1 PMID:29186529_1 PMID:29186529_2 PMID:29186529_3 PMID:28789705_1 
#>              17               2              34               2               5 
#> PMID:28789705_2 PMID:28789705_3 PMID:25444008_1 PMID:25444008_2 PMID:25444008_3 
#>               2               1               7               3               3 
#> PMID:25444008_4 PMID:25444008_5 PMID:32012716_1 PMID:32012716_2 PMID:31006170_1 
#>               3               1               2               2               4 
#> PMID:31006170_2 PMID:31006170_3 PMID:31006170_4 PMID:32380969_1 PMID:32380969_2 
#>               4               2               1               1               2 
#> PMID:31337807_1 PMID:31337807_2 PMID:31337807_3 PMID:31337807_4 PMID:29554907_1 
#>               1               1               1               1               3 
#> PMID:29554907_2 PMID:28808302_1 PMID:28808302_2 PMID:28808302_3 PMID:28808302_4 
#>               1               4               1               1               2 
#> PMID:31226994_1 PMID:31226994_2 PMID:31976177_1 PMID:31976177_2 PMID:30568265_1 
#>               1               4               5              43               1 
#> PMID:30568265_2 PMID:30568265_3 PMID:30366118_1 PMID:30267022_1 PMID:26551842_1 
#>               1               3               1              10               2 
#> PMID:26551842_2 PMID:26551842_3 PMID:26551842_4 PMID:32143621_1 PMID:32143621_2 
#>               2               1               1               4              15 
#> PMID:31934770_1 PMID:19849869_1 PMID:22546742_1 PMID:23631345_1 PMID:32429742_1 
#>               1               1               1               1               1 
#> PMID:25761741_1 PMID:25761741_2 PMID:30534614_1 PMID:30534614_2 PMID:23459324_1 
#>               6               6               4               4               1 
#> PMID:23459324_2 PMID:29935270_1 PMID:29935270_2 
#>               1               1               4 

getCriticalN(relevant.sigs, my.siglengths)

#> 95% 
#>  10 

# Compare to observed
frequencySigs(my.sigs.increased)

#>           f__Enterobacteriaceae                g__Lactobacillus 
#>                               8                               6 
#>            s__Lactobacillus sp.              f__Lachnospiraceae 
#>                               5                               5 
#>               f__Micrococcaceae                  g__Bacteroides 
#>                               4                               4 
#>                   p__Firmicutes                  g__Clostridium 
#>                               4                               4 
#> s__Faecalibacterium prausnitzii           s__Veillonella dispar 
#>                               4                               4